Min requirements (#307)
* ..

* apply the click directive on some components

* chang bg opacity

* fixed game history

* decrease paddle speed

* fix invite issue

* change well known components

* bug fix

* use sm, arg in custom directive, fix sound track

* apply sound effect on all game components

* fix click sound effect

* add main theme ost

* fix

* fix theme ost

* game score don't change when game end by leave or giveup

* sound button in home page

* fix audio toggle in play game

* fixed seeder

* fix button a bit

* fixed controller name

* fixed match history

* fixed match history

* Add MainPopup component and use it for adding users to group chat

* cleaned the seeders and changed usernames

* fixed match

* split create group chat into smaller reusable component

* format files

* rename Game Gateway files

* merged

* allowed only the owner to change group settings

* formatted

* #170, add dimmed background for both chat & friends modals

* fixed quiries

* adding click effect directive in chat components

* allow owner only to change chat Type

* remove console logs

* allow owner onlt to chat chat info

* format files

* fix conflicts

* add some style to chat info, format files

* rename confusing arguments, properties etc..

* more renames

* Computer is always ready

* fix game invite bug, game not starting sometimes

* add game invite to help page

* add gameinvite to  layout

* clean

* add title to show status on hover

* update friends status live

* fixed owner leave

* fixed username crash

* fix after conflicts

* remove dots from game selector

* fix the mute button

* fix from username to userlogin in stats

* ..

* change from username to login frontend

* change playerID to lgin

* change player to login in gameanalayzer

* just to push

* paunish setup

* paunish leaver

* Change username to login in PlayerDto

* update UserProfileSetup theme

* use login instead of username in friends

* paunish does work

* fix but, decline not working

* fix compensate & paunish players

* improve popup

* clean up

* fix bugs after merge

* enable rotation + resize in pc only

* reset camera position to default using Space key

* bug fix

* fixed pages

* added username regex

* improve announce popup to handle paunish and compensate announce

* fix notification to retrieve the correct type of notification

* better code

* done with announcing

* update leaderboard

* set min zoom in, zoom out. and disable pan

* fix merge issue

* remove logs

* adding deuce

* fix one issue with annonce

* fix order of announce

* :D

* fix warning

* adding deuce event

* fix emit deuce from back-end

* fix condition of emitting the deuce

* clean test vars

* Handle responsivity for Home page

* temp page style for help page

* fix merge issues

* all in 1

* styles is done

* almost done

* clean the code

* #247

* images only remains

* fix bugs

* fix minor bugs m

* update theme and responsiveness of sort popover in achievement component

* fix style issues

* fix all bugs with popup transition

* #248

* fix matches appear in history if page was refreshed

* friend error fix m

* #258

* Handle pagination for chat list, format files

* removed pass

* make loading component, add component for default layout and callback page

* fixed pass issue

* fix transition error, fix chat & friend modals opacity bug

* groupt chat limited

* change size of loading component and make it resizable

* limited name to 30

* remove condition from isProtait

* change approach to detect isMoblie

* fix style in confirm page

* format files

* fix timer bug in confirm page

* #271

* Update Game colors to match theme

* Modify Game Components Style fix #256

* clean

* increase ball speed each hit (classic only)

* fix orinitation detect

* .

* add background for the button

* update

* change email name

* push changes for real now

* add imgs to achv

* add imgs to achv

* achv img is done

* clean the cpde

* replace @click by v-click-effect

* remove logs

* arrange some buttons

* Delete app/web/components/game directory

* add img for rank system only their names remains

* 277

* update notification message on failed update for username and image

* 292

* create common SideBar component for both chat and friend modals

* #282

* disconnect from gateways on logout, fix #281

* 295

* used username instead

* #282

* #278

* #289

* #282

* format files

* fix #285 #287

* .

* ..

* fixed another bug in invite

* add ranks to help page

* add another check in respond-invite

* fix Remove 'v-click-effect' #297

* fix issue leaderboard is ordered by matches not wins #283

* update and close Remove 'v-click-effect' #297

* fix notifications with socket exceptions

* update notification message

* update

* fix & close  Remove 'v-click-effect' #297

* print payload.content for notification message

* fix primevue errors

* wallahi it is moatasem

* wallahi it is moatasem

* update error messages for friends notification

* handle socket protection

* #297

* remove logs on frontend

* fix primevue issue

---------

+++ app/api/src/module/friend/friend.controller.ts
@@ -1,31 +1,44 @@
 import { FriendService } from './friend.service'
-import { UserGetDto } from './dto/friend.dto'
-import { Get, Controller, Param, Post, Delete } from '@nestjs/common'
+import { Get, Controller, Param, Post, Delete, BadRequestException } from '@nestjs/common'
 import { JwtAuthGuard } from '../../common/guards/jwt.guard'
 import { UseGuards, Req } from '@nestjs/common'
+import { UserGetDto } from '@module/user/dto/user.dto'
+import { ParseStringPipe } from '@common/pipes/string.pipe'
+@UseGuards(JwtAuthGuard)
 @Controller('/users/:user/friends')
 export class FriendController {
     constructor(private readonly FriendService: FriendService) {}
 
+    @Get('')
+    async GetFriends(@Req() req): Promise<UserGetDto[]> {
+        return await this.FriendService.getFriends(req.user.login)
+    }
+
     @Post('/:friend')
     async UpdateFriend(
-        @Param('user') user: string,
-        @Param('friend') friend: string,
+        @Param('user', ParseStringPipe) user: string,
+        @Param('friend', ParseStringPipe) friend: string,
+        @Req() req,
     ): Promise<UserGetDto> {
-        return await this.FriendService.CheckFriendsUpdate(user, friend)
+        if (user === friend) throw new BadRequestException('You cannot add yourself as a friend')
+        if (user !== req.user.login)
+            throw new BadRequestException('You cannot add a friend for someone else')
+        const newFriend = await this.FriendService.CheckFriendsUpdate(user, friend)
+        if (!newFriend) throw new BadRequestException('Friend not found')
+        return newFriend
     }
 
     @Delete('/:friend')
     async DeleteFriend(
-        @Param('user') user: string,
-        @Param('friend') friend: string,
+        @Param('user', ParseStringPipe) user: string,
+        @Param('friend', ParseStringPipe) friend: string,
+        @Req() req,
     ): Promise<UserGetDto> {
-        return await this.FriendService.DeleteFriend(friend, user)
-    }
-
-    @UseGuards(JwtAuthGuard)
-    @Get('')
-    async GetFriends(@Req() req): Promise<UserGetDto[]> {
-        return await this.FriendService.getFriends(req.user.login)
+        if (user === friend) throw new BadRequestException('You cannot remove yourself as a friend')
+        if (user !== req.user.login)
+            throw new BadRequestException('You cannot remove a friend for someone else')
+        const deletedFriend = await this.FriendService.DeleteFriend(friend, user)
+        if (!deletedFriend) throw new BadRequestException('Friend not found')
+        return deletedFriend
     }
 }
